requirements.txt
django>=3.0.1
psycopg2-binary>=2.8.4
djangorestframework>=3.11.0
Pillow>=7.0.0
BeautifulSoup4>=4.8.2
requests>=2.23